perf: leverage jsx for settings, easier onboading
