+++ cypress/.eslintrc.yaml
@@ -1,2 +1,6 @@
 globals:
   cy: readonly
+  it: readonly
+  describe: readonly
+  beforeEach: readonly
+  expect: readonly
