+++ package.json
@@ -22,10 +22,13 @@
     "@nuxtjs/i18n": "^7.3.0",
     "@nuxtjs/pwa": "^3.3.5",
     "@nuxtjs/sentry": "^6.0.0",
+    "bcrypt": "^5.0.1",
     "bootstrap-vue": "^2.22.0",
+    "cheerio": "^1.0.0-rc.12",
     "core-js": "^3.25.1",
     "dayjs": "^1.11.5",
     "leaflet": "1.3.1",
+    "node-fetch": "^2.6.7",
     "nuxt": "^2.15.8",
     "vue": "^2.7.10",
     "vue-binpacker-plugin": "^1.1.0",
@@ -44,10 +47,12 @@
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.18.9",
+    "@babel/runtime-corejs3": "^7.19.0",
     "@cypress/vue": "^2.2.4",
     "@cypress/webpack-dev-server": "^2.3.0",
+    "@nuxt/babel-preset-app": "^2.15.8",
     "@nuxt/types": "~2.15.0",
-    "@nuxtjs/eslint-config": "^9.0.0",
+    "@nuxtjs/eslint-config": "^11.0.0",
     "@nuxtjs/eslint-module": "^3.1.0",
     "cypress": "^10.7.0",
     "eslint": "^8.4.1",
@@ -80,7 +85,12 @@
     },
     "parserOptions": {
       "parser": "@babel/eslint-parser",
-      "requireConfigFile": false
+      "requireConfigFile": false,
+      "babelOptions": {
+        "presets": [
+          "@nuxt/babel-preset-app"
+        ]
+      }
     },
     "extends": [
       "@nuxtjs",
